Removed lint warning on poggit.
