+++ src/matcracker/BlocksConverter/Loader.php
@@ -7,11 +7,10 @@ namespace matcracker\BlocksConverter;
 use matcracker\BlocksConverter\commands\Convert;
 use matcracker\BlocksConverter\commands\ConvertQueue;
 use matcracker\BlocksConverter\commands\ToolBlock;
-use pocketmine\block\BlockIds;
+use matcracker\BlocksConverter\tasks\ToolBlockTask;
 use pocketmine\event\Listener;
 use pocketmine\event\player\PlayerQuitEvent;
 use pocketmine\plugin\PluginBase;
-use pocketmine\scheduler\Task;
 
 final class Loader extends PluginBase implements Listener
 {
@@ -29,21 +28,7 @@ final class Loader extends PluginBase implements Listener
 		$this->getServer()->getCommandMap()->register('convertqueue', new ConvertQueue($this));
 		$this->getServer()->getCommandMap()->register('toolblock', new ToolBlock());
 
-		$this->getScheduler()->scheduleRepeatingTask(new class extends Task
-		{
-			public function onRun(int $currentTick)
-			{
-				$players = ToolBlock::getPlayers();
-				foreach ($players as $player) {
-					$block = $player->getTargetBlock(5);
-					if ($block !== null && $block->getId() !== BlockIds::AIR) {
-						$message = "{$block->getName()} (ID: {$block->getId()} Meta: {$block->getDamage()})\n";
-						$message .= "X: {$block->getX()} Y: {$block->getY()} Z: {$block->getZ()}";
-						$player->sendTip($message);
-					}
-				}
-			}
-		}, 5);
+		$this->getScheduler()->scheduleRepeatingTask(new ToolBlockTask(), 5);
 	}
 
 	public function onPlayerQuit(PlayerQuitEvent $event): void
